updated concat with correct tensors.
In order to allow us to inspect the tensor sizes, it was nessacary to allow the graph to build. So we hacked the concat operation and used identical tensors. Now that the size of up_conv_prev_layer,down_layer match, we use them for concat.
The graph builds and runs. dims_up[1] shows the size of the tensor at the last up layer.
Now the issue is that the size is not the same as input image.